requirements.txt
google-generativeai 
datasets
litellm 
pydantic